Somewhat rethink how attribute selection should work
Do a general [attributeName operator value] instead. Attributename should maybe be a dropdown list of functions AND previously(?) defined variables. Operators are just boolean operators as well as in. in should only be able to operate on a list. Then the value is either a literal, number value such as 4, or a name of some part such as "facade"? Idk.
Here's an idea, maybe the attributeName is what determines the things that are shown. Howaboutit?